Renamed package (again) to dnf-plugin-cow
Using sub-package for python3 based DNF plugin. The goal is to
eventually replace this with libdnf-plugin-cow. See
https://github.com/facebookincubator/dnf-plugin-cow/issues/1
for rationale

Use Recommends instead of Requires to allow the package to be installed
without complementary changes for rpm and librepo. This is intended to
be temporary until Fedora Rawhide includes the changes.

Fixed the date in an old changelog. Probably was working past midnight
and got the day of week/date mixed up.